Consolidate user role handling for API key updates
