+++ x-pack/plugin/security/src/main/java/org/elasticsearch/xpack/security/action/apikey/TransportGrantApiKeyAction.java
@@ -21,7 +21,7 @@ import org.elasticsearch.xpack.core.security.action.apikey.GrantApiKeyRequest;
 import org.elasticsearch.xpack.security.action.TransportGrantAction;
 import org.elasticsearch.xpack.security.authc.ApiKeyService;
 import org.elasticsearch.xpack.security.authc.AuthenticationService;
-import org.elasticsearch.xpack.security.authc.support.ApiKeyManager;
+import org.elasticsearch.xpack.security.authc.support.ApiKeyGenerator;
 import org.elasticsearch.xpack.security.authz.AuthorizationService;
 import org.elasticsearch.xpack.security.authz.store.CompositeRolesStore;
 
@@ -30,7 +30,7 @@ import org.elasticsearch.xpack.security.authz.store.CompositeRolesStore;
  */
 public final class TransportGrantApiKeyAction extends TransportGrantAction<GrantApiKeyRequest, CreateApiKeyResponse> {
 
-    private final ApiKeyManager apiKeyManager;
+    private final ApiKeyGenerator generator;
 
     @Inject
     public TransportGrantApiKeyAction(
@@ -47,7 +47,7 @@ public final class TransportGrantApiKeyAction extends TransportGrantAction<Grant
             transportService,
             actionFilters,
             threadPool.getThreadContext(),
-            new ApiKeyManager(apiKeyService, rolesStore, xContentRegistry),
+            new ApiKeyGenerator(apiKeyService, rolesStore, xContentRegistry),
             authenticationService,
             authorizationService
         );
@@ -58,7 +58,7 @@ public final class TransportGrantApiKeyAction extends TransportGrantAction<Grant
         TransportService transportService,
         ActionFilters actionFilters,
         ThreadContext threadContext,
-        ApiKeyManager apiKeyManager,
+        ApiKeyGenerator generator,
         AuthenticationService authenticationService,
         AuthorizationService authorizationService
     ) {
@@ -71,16 +71,14 @@ public final class TransportGrantApiKeyAction extends TransportGrantAction<Grant
             authorizationService,
             threadContext
         );
-        this.apiKeyManager = apiKeyManager;
+        this.generator = generator;
     }
 
     @Override
     protected void doExecute(Task task, GrantApiKeyRequest request, ActionListener<CreateApiKeyResponse> listener) {
         executeWithGrantAuthentication(
             request,
-            listener.delegateFailure(
-                (l, authentication) -> apiKeyManager.generateApiKey(authentication, request.getApiKeyRequest(), listener)
-            )
+            listener.delegateFailure((l, authentication) -> generator.generateApiKey(authentication, request.getApiKeyRequest(), listener))
         );
     }
 }
