+++ x-pack/plugin/security/src/test/java/org/elasticsearch/xpack/security/authc/support/ApiKeyUserRoleDescriptorResolverTests.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+package org.elasticsearch.xpack.security.authc.support;
+
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.support.PlainActionFuture;
+import org.elasticsearch.common.util.set.Sets;
+import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.xcontent.NamedXContentRegistry;
+import org.elasticsearch.xpack.core.security.authc.Authentication;
+import org.elasticsearch.xpack.core.security.authc.AuthenticationTestHelper;
+import org.elasticsearch.xpack.core.security.authc.Subject;
+import org.elasticsearch.xpack.core.security.authz.RoleDescriptor;
+import org.elasticsearch.xpack.core.security.user.User;
+import org.elasticsearch.xpack.security.authz.store.CompositeRolesStore;
+import org.junit.Test;
+
+import java.util.Collection;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static org.hamcrest.Matchers.arrayWithSize;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.mock;
+
+public class ApiKeyUserRoleDescriptorResolverTests extends ESTestCase {
+
+    @SuppressWarnings("unchecked")
+    @Test
+    public void testGetRoleDescriptors() {
+        final CompositeRolesStore rolesStore = mock(CompositeRolesStore.class);
+        final ApiKeyUserRoleDescriptorResolver resolver = new ApiKeyUserRoleDescriptorResolver(rolesStore, NamedXContentRegistry.EMPTY);
+        final Set<String> userRoleNames = Sets.newHashSet(randomArray(1, 4, String[]::new, () -> randomAlphaOfLengthBetween(3, 12)));
+        final Authentication authentication = AuthenticationTestHelper.builder()
+            .user(new User("test", userRoleNames.toArray(String[]::new)))
+            .realmRef(new Authentication.RealmRef("realm-name", "realm-type", "node-name"))
+            .build(false);
+        final Set<RoleDescriptor> roleDescriptors = randomSubsetOf(userRoleNames).stream()
+            .map(name -> new RoleDescriptor(name, generateRandomStringArray(3, 6, false), null, null))
+            .collect(Collectors.toUnmodifiableSet());
+
+        doAnswer(inv1 -> {
+            final Object[] args1 = inv1.getArguments();
+            assertThat(args1, arrayWithSize(2));
+
+            Subject subject = (Subject) args1[0];
+            assertThat(subject.getType(), is(Subject.Type.USER));
+            assertThat(Set.of(subject.getUser().roles()), equalTo(userRoleNames));
+
+            ActionListener<Collection<Set<RoleDescriptor>>> listener1 = (ActionListener<Collection<Set<RoleDescriptor>>>) args1[args1.length
+                - 1];
+            listener1.onResponse(List.of(roleDescriptors));
+            return null;
+        }).when(rolesStore).getRoleDescriptorsList(any(Subject.class), any(ActionListener.class));
+
+        final PlainActionFuture<Set<RoleDescriptor>> future = new PlainActionFuture<>();
+        resolver.getUserRoleDescriptors(authentication, future);
+
+        assertThat(future.actionGet(), equalTo(roleDescriptors));
+    }
+}
